replace io/blocking with io/ws, this is the way to go
