+++ io/blocking/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
-name = "magic-wormhole-io-blocking"
+name = "magic-wormhole-io-ws"
 version = "0.0.1"
 authors = ["Brian Warner <warner@lothar.com>"]
 
 [dependencies]
-websocket = "0.20"
 magic-wormhole-core = { version = "0.0.1", path = "../../core" }
-futures = "0.1"
-tokio-core = "0.1"
+ws = "0.7.*"
+url = "1.7"
 
+[dev-dependencies]
+hex = "0.3"
