+++ io/blocking/examples/send.rs
@@ -1,5 +1,6 @@
-extern crate magic_wormhole_io_blocking;
-use magic_wormhole_io_blocking::Wormhole;
+extern crate magic_wormhole_io_ws;
+extern crate hex;
+use magic_wormhole_io_ws::Wormhole;
 
 // Can ws do hostname lookup? Use ip addr, not localhost, for now
 const MAILBOX_SERVER: &'static str = "ws://127.0.0.1:4000/v1";
@@ -11,7 +12,13 @@ fn main() {
     w.set_code("4-purple-sausages");
     println!("sending..");
     w.send_message(b"hello");
-    println!("sent, closing..");
+    println!("sent..");
+    // if we close right away, we won't actually send anything. Wait for at
+    // least the verifier to be printed, that ought to give our outbound
+    // message a chance to be delivered.
+    let verifier = w.get_verifier();
+    println!("verifier: {}", hex::encode(verifier));
+    println!("got verifier, closing..");
     w.close();
     println!("closed");
 }
