--- io/ws/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "magic-wormhole-io-ws"
-version = "0.0.1"
-authors = ["Brian Warner <warner@lothar.com>"]
-
-[dependencies]
-magic-wormhole-core = { version = "0.0.1", path = "../../core" }
-ws = "0.7.*"
-url = "1.7"
-
-[dev-dependencies]
-hex = "0.3"
